+++ HISTORY.md
@@ -2,6 +2,7 @@
 -----------------
 - Bumped mongodb-core to 1.2.23.
 - Added Cursor.prototype.setCursorOption to allow for setting node specific cursor options for tailable cursors.
+- NODE-430 Cursor.count() opts argument masked by var opts = {}
 
 2.0.27 04-07-2015
 -----------------